rename string.ts to capitalizeFirstLetter
