+++ frontend/src/utils/capitalizeFirstLetter.ts
